requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0
python-multipart>=0.0.6
torch --index-url https://download.pytorch.org/whl/cpu
transformers
Pillow
